+++ packages/mongo_document_annotation/example/lib/models/comment.mongo_document.dart
@@ -291,7 +291,7 @@ class Comments {
 
     final selector = <String, dynamic>{};
     if (id != null) selector['_id'] = id;
-    if (post != null) selector['post'] = post;
+    if (post != null) selector['post'] = post.id;
     if (text != null) selector['text'] = text;
     if (age != null) selector['age'] = age;
     if (createdAt != null) selector['created_at'] = createdAt;
@@ -411,7 +411,7 @@ class Comments {
 
     final selector = <String, dynamic>{};
     if (id != null) selector['_id'] = id;
-    if (post != null) selector['post'] = post;
+    if (post != null) selector['post'] = post.id;
     if (text != null) selector['text'] = text;
     if (age != null) selector['age'] = age;
     if (createdAt != null) selector['created_at'] = createdAt;