+++ packages/mongo_document_generator/lib/src/templates/crud/read/read_templates.dart
@@ -83,6 +83,7 @@ class ReadTemplates {
     FieldRename? fieldRename,
     List<ParameterElement> params,
     String className,
+    Map<String, dynamic> nestedCollectionMap,
   ) {
     String classNameVar = className.toLowerCase();
     return '''
@@ -96,7 +97,7 @@ class ReadTemplates {
       final paramName = p.name;
       final key =
           ParameterTemplates.getParameterKey(typeChecker, p, fieldRename);
-      return '''if ($paramName != null) selector['$key'] = $paramName;''';
+      return '''if ($paramName != null) selector['$key'] = ${nestedCollectionMap.containsKey(paramName) ? "$paramName.id" : paramName};''';
     }).join('\n')}
     if (selector.isEmpty) {
       final $classNameVar = await coll.modernFindOne(sort: {'created_at': -1});
@@ -160,6 +161,7 @@ class ReadTemplates {
     FieldRename? fieldRename,
     List<ParameterElement> params,
     String className,
+    Map<String, dynamic> nestedCollectionMap,
   ) {
     String classNameVar = className.toLowerCase();
     return '''
@@ -173,7 +175,7 @@ class ReadTemplates {
       final paramName = p.name;
       final key =
           ParameterTemplates.getParameterKey(typeChecker, p, fieldRename);
-      return '''if ($paramName != null) selector['$key'] = $paramName;''';
+      return '''if ($paramName != null) selector['$key'] = ${nestedCollectionMap.containsKey(paramName) ? "$paramName.id" : paramName};''';
     }).join('\n')}
     if (selector.isEmpty) {
       final ${classNameVar}s = await coll.modernFind(sort: {'created_at': -1},limit:limit,skip:skip).toList();